+++ src/prompt.c
@@ -18,7 +18,7 @@ void format_directory(char **directory)
     }
     else
     {
-        for (int i = 0; i < homelen; i++)
+        for (size_t i = 0; i < homelen; i++)
         {
             if (home[i] != (*directory)[i])
             {
